Move NodeRegister to dev API To fix LTO issue for eye decomposition
